+++ front/e2e/src/po/deconnexion.po.ts
@@ -1,0 +1,9 @@
+import { $ } from 'protractor';
+import { AppPage } from './app.po';
+
+export class DeconnexionPage extends AppPage {
+
+  async cliquerSurSeReconnecter() {
+    await $('#btnSeReconnecter').click();
+  }
+}